+++ new-client/src/plugins/Sketch/components/featureStyle/FeatureStyleSelector.js
@@ -5,6 +5,7 @@ import { STROKE_DASHES, DEFAULT_DRAW_STYLE_SETTINGS } from "../../constants";
 import FeatureStyleAccordion from "./FeatureStyleAccordion";
 import FeaturePointSizeAccordion from "./FeatureSizeAccordion";
 import StrokeTypeSelector from "./StrokeTypeSelector";
+import FeatureBufferAccordion from "./FeatureBufferAccordion";
 
 export default function FeatureStyleSelector(props) {
   const { activeDrawType, drawStyle, setDrawStyle } = props;
@@ -177,6 +178,10 @@ export default function FeatureStyleSelector(props) {
   };
 
   const renderPointStyleSettings = () => {
+    // If the isBufferStyle is true, we don't want to show the point-size-slider on edit.
+    if (props.bufferState.isBufferStyle) {
+      return null;
+    }
     return (
       <Grid container>
         <Grid item xs={12}>
@@ -244,6 +249,25 @@ export default function FeatureStyleSelector(props) {
     );
   };
 
+  const renderBufferStyleSettings = () => {
+    return (
+      <FeatureBufferAccordion
+        title="Buffra"
+        showBufferSlider
+        drawStyle={props.drawStyle}
+        drawModel={props.drawModel}
+        localObserver={props.localObserver}
+        globalObserver={props.globalObserver}
+        pluginShown={props.pluginShown}
+        bufferState={props.bufferState}
+        setBufferState={props.setBufferState}
+        highlightLayer={props.highlightLayer}
+        toggleBufferBtn={props.toggleBufferBtn}
+        setToggleBufferBtn={props.setToggleBufferBtn}
+      />
+    );
+  };
+
   // We want to display different settings depending on what the user is drawing!
   // Let's check and render the appropriate settings.
   const renderColorSelectors = () => {
@@ -296,6 +320,9 @@ export default function FeatureStyleSelector(props) {
         <Grid item xs={12}>
           {renderColorSelectors()}
           {props.activeDrawType === "Point" && renderPointStyleSettings()}
+          {props.activityId === "ADD" &&
+            props.activeDrawType !== "Circle" &&
+            renderBufferStyleSettings()}
         </Grid>
       </Grid>
     </Grid>